+++ src/models/reporting.ts
@@ -16,57 +16,13 @@
  */
 // eslint-disable-next-line import/named
 import { RunnerTask, TaskMeta } from 'vitest';
-import { Attribute, Issue } from './common';
-import { TEST_ITEM_TYPES, LOG_LEVELS, LAUNCH_MODES } from '../constants';
-
-export interface StartLaunchObjType {
-  startTime?: string | number;
-  attributes?: Array<Attribute>;
-  description?: string;
-  name?: string;
-  rerun?: boolean;
-  rerunOf?: string;
-  mode?: LAUNCH_MODES;
-  id?: string;
-}
-
-export interface StartTestObjType {
-  name: string;
-  type: TEST_ITEM_TYPES;
-  attributes?: Array<Attribute>;
-  description?: string;
-  startTime?: string | number;
-  codeRef?: string;
-  testCaseId?: string;
-  retry?: boolean;
-}
-
-export interface FinishTestItemObjType {
-  endTime?: string | number;
-  status?: string;
-  attributes?: Attribute[];
-  description?: string;
-  testCaseId?: string;
-  issue?: Issue;
-}
-
-export interface Attachment {
-  name: string;
-  type: string;
-  content: string | Buffer;
-}
-
-export interface LogRQ {
-  level?: LOG_LEVELS;
-  message?: string;
-  time?: string | number;
-  file?: Attachment;
-}
+import type { LogOptions, Attachment, Attribute } from '@reportportal/client-javascript/models';
+import { LOG_LEVELS } from '../constants';
 
 export interface RPTaskMeta extends TaskMeta {
   rpMeta: {
     test: {
-      logs: LogRQ[];
+      logs: LogOptions[];
       attributes: Attribute[];
       testCaseId?: string;
       description?: string;